Moved from/to vector to tensor.hpp
