+++ ttnn/cpp/ttnn/tensor/CMakeLists.txt
@@ -11,7 +11,6 @@ set(TENSOR_SRCS
     ${CMAKE_CURRENT_SOURCE_DIR}/layout/page_config.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/layout/size.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/layout/tensor_layout.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/xtensor/conversion_utils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/xtensor/partition.cpp
     CACHE INTERNAL
     "Tensor sources to reuse in ttnn build"
